requirements.txt
evaluate
tensorflow_hub
datasets
pytest
hydra-core
multiprocess
numpy
omegaconf
pandas
peft
pyserini
pytrec_eval
regex
rouge
scikit_learn
scipy
tensorflow==2.8.0
protobuf==3.20.*
tensorflow_hub
torch
tqdm
transformers
trl
bitsandbytes
tabulate
wandb
pytest
flash_attn